Title: 앞에서부터 삭제하기 2, Time: 2483ms, Memory: 22MB, Status: Passed - CodeTree

+++ 250204/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -27,7 +27,7 @@ public class Main {
             sum -= arr[i];
             pq.remove(arr[i]);
 
-            int avg = (sum - pq.peek()) / (pq.size() - 1);
+            double avg = (double) (sum - pq.peek()) / (pq.size() - 1);
 
             max = Math.max(avg, max);
         }
